# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is an API documentation for SriRejeki IoT Greenhouse Monitoring",
        "title": "SriRejeki Backend API Documentation",
        "contact": {
            "name": "API Support",
            "email": "mail@hilmy.dev"
        },
        "version": "1.0"
    },
    "host": "localhost:80",
    "basePath": "/",
    "paths": {
        "/auth": {
            "get": {
                "description": "Get user session data using authorization bearer header",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Get user session data",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/auth.authResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/auth.authResponse"
                        }
                    }
                }
            }
        },
...
        "/farmland": {
            "get": {
                "description": "Get one farmlands data using authorization bearer header",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Farmland"
                ],
                "summary": "Get one farmlands data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Farmland Id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/farmland.farmlandResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/farmland.farmlandResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/farmland.farmlandResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/farmland.farmlandResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Add new farmland data using authorization bearer header",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Farmland"
                ],
                "summary": "Add new farmland data",
                "parameters": [
                    {
                        "description": "Required data to add new farmland",
                        "name": "farmlandAddRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/farmland.farmlandAddRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/farmland.farmlandResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/farmland.farmlandResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/farmland.farmlandResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/farmland.farmlandResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete existing farmland data using authorization bearer header",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Farmland"
                ],
                "summary": "Delete existing farmland data",
                "parameters": [
                    {
                        "description": "Required data to delete existing farmland",
                        "name": "farmlandDeleteRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/farmland.farmlandDeleteRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/farmland.farmlandResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/farmland.farmlandResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/farmland.farmlandResponse"
                        }
                    }
                }
            },
            "patch": {
                "description": "Edit existing farmland data using authorization bearer header",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Farmland"
                ],
                "summary": "Edit existing farmland data",
                "parameters": [
                    {
                        "description": "Required data to edit existing farmland",
                        "name": "farmlandEditRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/farmland.farmlandEditRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/farmland.farmlandResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/farmland.farmlandResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/farmland.farmlandResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/farmland.farmlandResponse"
                        }
                    }
                }
            }
        },
        "/farmland/user": {
            "get": {
                "description": "Get all user's farmlands data using authorization bearer header",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Farmland"
                ],
                "summary": "Get all user's farmlands data",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/farmland.farmlandResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/farmland.farmlandResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/farmland.farmlandResponse"
                        }
                    }
                }
            }
        },
...
        "/mcu": {
            "get": {
                "description": "Get one microcontroller data using authorization bearer header",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Microcontroller"
                ],
                "summary": "Get one microcontroller data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Microcontroller Id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mcu.microcontrollerResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/mcu.microcontrollerResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/mcu.microcontrollerResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/mcu.microcontrollerResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Add new microcontroller data using authorization bearer header",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Microcontroller"
                ],
                "summary": "Add new microcontroller data",
                "parameters": [
                    {
                        "description": "Required data to add new microcontroller",
                        "name": "microcontrollerAddRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mcu.microcontrollerAddRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/mcu.microcontrollerResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/mcu.microcontrollerResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/mcu.microcontrollerResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/mcu.microcontrollerResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete existing microcontroller data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Microcontroller"
                ],
                "summary": "Delete existing microcontroller data",
                "parameters": [
                    {
                        "description": "Required data to delete existing microcontroller",
                        "name": "microcontrollerDeleteRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mcu.microcontrollerDeleteRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mcu.microcontrollerResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/mcu.microcontrollerResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/mcu.microcontrollerResponse"
                        }
                    }
                }
            },
            "patch": {
                "description": "Edit existing microcontroller data using authorization bearer header",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Microcontroller"
                ],
                "summary": "Edit existing microcontroller data",
                "parameters": [
                    {
                        "description": "Required data to edit existing microcontroller",
                        "name": "microcontrollerEditRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mcu.microcontrollerEditRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mcu.microcontrollerResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/mcu.microcontrollerResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/mcu.microcontrollerResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/mcu.microcontrollerResponse"
                        }
                    }
                }
            }
        },
        "/mcu/auth": {
            "post": {
                "description": "Authenticate microcontroller | INTERNAL USE ON BACKEND ONLY",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Microcontroller"
                ],
                "summary": "Authenticate microcontroller | INTERNAL USE ON BACKEND ONLY",
                "parameters": [
                    {
                        "description": "Required data to authenticate microcontroller",
                        "name": "microcontrollerAuthRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mcu.microcontrollerAuthRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mcu.microcontrollerResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/mcu.microcontrollerResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/mcu.microcontrollerResponse"
                        }
                    }
                }
            }
        },
        "/mcu/farmland": {
            "get": {
                "description": "Get all farmland's microcontrollers data using authorization bearer header",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Microcontroller"
                ],
                "summary": "Get all farmland's microcontrollers data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Farmland Id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mcu.microcontrollerResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/mcu.microcontrollerResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/mcu.microcontrollerResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/mcu.microcontrollerResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "auth.authResponse": {
            "type": "object",
            "properties": {
                "payload": {}
            }
        },
...
        "farmland.farmlandAddRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "farmland.farmlandDeleteRequest": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
        "farmland.farmlandEditRequest": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "location": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "farmland.farmlandResponse": {
            "type": "object",
            "properties": {
                "payload": {}
            }
        },
        "mcu.microcontrollerAddRequest": {
            "type": "object",
            "required": [
                "deviceId",
                "farmlandId",
                "name"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "deviceId": {
                    "type": "string"
                },
                "farmlandId": {
                    "type": "integer"
                },
                "location": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "mcu.microcontrollerResponse": {
            "type": "object",
            "properties": {
                "payload": {}
            }
        },
...
    }
}